feat: add support for dlq metrics
The current metrics reporting reported error even if a handler failed,
which might happen in case we receive the message faster than it is
committed to the database.

This patch ensures that we he warning level logging for handlers and
detailed stats, but also we measure the whole e2e flow.